requirements.txt
django-jquery==1.12.2 
simplejson==3.8.2 
six==1.10.0
